Add Ego3DVQA pipeline: frame selection, preprocessing, training, eval, and quality analysis
Adapt LangSplat for egocentric Ego3DVQA datasets (ADT, HD-EPIC) with covisibility-based
frame selection, mask postprocessing, and qualitative evaluation. Documents root causes
of poor segmentation quality on complex indoor egocentric scenes (autoencoder 3D bottleneck,
low Gaussian count, scene complexity).

New scripts:
- select_frames.py: covisibility + blur frame selection
- prepare_ego3dvqa_workspace.py: 2x downscale + workspace setup
- postprocess_segmaps.py: apply hand/object masks to seg maps
- visualize_preprocessing.py: diagnostic montages
- eval_ego3dvqa.py: CLIP relevancy heatmap evaluation
- run_ego3dvqa_pipeline.sh / _resume.sh: orchestration
- utils/covisibility.py: frame selection algorithm

Modified:
- preprocess.py: per-image processing with resume support (avoid OOM)
- scene/dataset_readers.py: skip missing images for subset datasets
- autoencoder/dataset.py: fp16->fp32 cast for CLIP features
- autoencoder/train.py: fix numpy/tensorboard histogram incompatibility

Docs:
- docs/pipeline/: pipeline execution report
- docs/experiments/: quality analysis with root cause diagnosis

+++ autoencoder/dataset.py
@@ -16,7 +16,7 @@ class Autoencoder_dataset(Dataset):
                 data = features
             else:
                 data = np.concatenate([data, features], axis=0)
-        self.data = data
+        self.data = data.astype(np.float32)  # CLIP outputs fp16, autoencoder needs fp32
 
     def __getitem__(self, index):
         data = torch.tensor(self.data[index])
